Settings Button, load Game State on Start Button Press
Added a settings button that when clicked allows user to see a drop down menu for difficulty selection ( Easy, Normal, Hard ) albeit, when selecting an option, nothing happens. When user selects the start game button, the game will load a simple scene for now just to make sure this is functional. I had the game load a blank screen that shows the money UI just to see if the button would function as expected.