MoneyHUD with Assets
Fleshed out the MoneyHUD to add visible assets. Made the container a bit nicer visually. Also added some simple scripts to game state, and moneyHUD that will listen for signals and we can test functionality by pressing A or C to raise values of the assets and cash. Wanted to have that to make sure the UI functions as expected. But they can be changed to what we agree on.